+++ R/reduce.R
@@ -1,9 +1,11 @@
-#' Reduce a list to a single value by iteratively applying a binary function.
+#' Reduce a list to a single value by iteratively applying a binary function
 #'
-#' `reduce()` combines from the left, `reduce_right()` combines from
-#' the right. `reduce(list(x1, x2, x3), f)` is equivalent to
-#' `f(f(x1, x2), x3)`; `reduce_right(list(x1, x2, x3), f)` is equivalent to
-#' `f(f(x3, x2), x1)`.
+#' @description
+#'
+#' `reduce()` is an operation that combines the elements of a vector
+#' into a single value. The combination is driven by `.f`, a binary
+#' function that takes two values and returns a single value: reducing
+#' `f` over `1:3` computes the value `f(f(1, 2), 3)`.
 #'
 #' @inheritParams map
 #' @param .y For `reduce2()`, an additional argument that is passed to
@@ -16,15 +18,62 @@
 #'   For `reduce2()`, a 3-argument function. The function will be passed the
 #'   accumulated value as the first argument, the next value of `.x` as the
 #'   second argument, and the next value of `.y` as the third argument.
-#'
 #' @param .init If supplied, will be used as the first value to start
-#'   the accumulation, rather than using \code{x[[1]]}. This is useful if
+#'   the accumulation, rather than using `x[[1]]`. This is useful if
 #'   you want to ensure that `reduce` returns a correct value when `.x`
 #'   is empty. If missing, and `x` is empty, will throw an error.
-#' @export
+#' @param .dir The direction of reduction as a string, one of `"left"`
+#'   or `"right"`. See the section about direction below.
+#'
+#' @section Direction:
+#'
+#' When `.f` is an associative operation like `+` or `c()`, the
+#' direction of reduction does not matter. For instance, reducing the
+#' vector `1:3` with the binary function `+` computes the sum `((1 +
+#' 2) + 3)` from the left, and the same sum `(1 + (2 + 3))` from the
+#' right.
+#'
+#' In other cases, the direction has important consequences on the
+#' reduced value. For instance, reducing a vector with `list()` from
+#' the left produces a left-leaning nested list (or tree), while
+#' reducing `list()` from the right produces a right-leaning list.
+#'
+#' @section Life cycle:
+#'
+#' `reduce_right()` is soft-deprecated as of purrr 0.3.0. Please use
+#' the `.dir` argument of `reduce()` instead. Note that the algorithm
+#' has changed. Whereas `reduce_right()` computed `f(f(3, 2), 1)`,
+#' `reduce(.dir = \"right\")` computes `f(1, f(2, 3))`. This is the
+#' standard way of reducing from the right.
+#'
+#' To update your code with the same reduction as `reduce_right()`,
+#' simply reverse your vector and use a left reduction:
+#'
+#' ```{r}
+#' # Before:
+#' reduce_right(1:3, f)
+#'
+#' # After:
+#' reduce(rev(1:3), f)
+#' ```
+#'
 #' @examples
+#' # Reducing `+` computes the sum of a vector while reducing `*`
+#' # computes the product:
 #' 1:3 %>% reduce(`+`)
 #' 1:10 %>% reduce(`*`)
+#'
+#' # When the operation is associative, the direction of reduction
+#' # does not matter:
+#' reduce(1:4, `+`)
+#' reduce(1:4, `+`, .dir = "right")
+#'
+#' # However with non-associative operations, the reduced value will
+#' # be different as a function of the direction. For instance,
+#' # `list()` will create left-leaning lists when reducing from the
+#' # right, and right-leaning lists otherwise:
+#' str(reduce(1:4, list))
+#' str(reduce(1:4, list, .dir = "right"))
 #'
 #' paste2 <- function(x, y, sep = ".") paste(x, y, sep = sep)
 #' letters[1:4] %>% reduce(paste2)
@@ -37,23 +86,14 @@
 #' reduce(samples, intersect)
 #'
 #' x <- list(c(0, 1), c(2, 3), c(4, 5))
-#' x %>% reduce(c)
-#' x %>% reduce_right(c)
-#' # Equivalent to:
-#' x %>% rev() %>% reduce(c)
-#'
 #' y <- list(c(6, 7), c(8, 9))
 #' reduce2(x, y, paste)
 #' reduce2_right(x, y, paste)
 #' # Equivalent to:
 #' x %>% rev() %>% reduce2(rev(y), paste)
-reduce <- function(.x, .f, ..., .init) {
-  reduce_impl(.x, .f, ..., .init = .init, .left = TRUE)
-}
-#' @rdname reduce
 #' @export
-reduce_right <- function(.x, .f, ..., .init) {
-  reduce_impl(.x, .f, ..., .init = .init, .left = FALSE)
+reduce <- function(.x, .f, ..., .init, .dir = c("left", "right")) {
+  reduce_impl(.x, .f, ..., .init = .init, .dir = .dir)
 }
 
 #' @rdname reduce
@@ -67,13 +107,25 @@ reduce2_right <- function(.x, .y, .f, ..., .init) {
   reduce2_impl(.x, .y, .f, ..., .init = .init, .left = FALSE)
 }
 
-reduce_impl <- function(.x, .f, ..., .init, .left = TRUE) {
-  out <- reduce_init(.x, .init, left = .left)
-  idx <- reduce_index(.x, .init, left = .left)
+reduce_impl <- function(.x, .f, ..., .init, .dir = "left") {
+  left <- arg_match(.dir, c("left", "right")) == "left"
+
+  out <- reduce_init(.x, .init, left = left)
+  idx <- reduce_index(.x, .init, left = left)
 
   .f <- as_mapper(.f, ...)
-  for (i in idx) {
-    out <- .f(out, .x[[i]], ...)
+
+  # Left-reduce produces left-leaning computation trees (reduced
+  # values are passed to the left) while right-reduce produces
+  # right-leaning trees
+  if (left) {
+    for (i in idx) {
+      out <- .f(out, .x[[i]], ...)
+    }
+  } else {
+    for (i in idx) {
+      out <- .f(.x[[i]], out, ...)
+    }
   }
 
   out
@@ -222,4 +274,34 @@ accumulate_names <- function(nms, init, right = FALSE) {
   }
 
   nms
+}
+
+#' Reduce from the right (retired)
+#'
+#' @description
+#'
+#' \Sexpr[results=rd, stage=render]{purrr:::lifecycle("soft-deprecated")}
+#'
+#' This function is retired as of purrr 0.3.0. Please use the `.dir`
+#' argument of [reduce()] instead.
+#'
+#' @inheritParams reduce
+#'
+#' @keywords internal
+#' @export
+reduce_right <- function(.x, .f, ..., .init) {
+  signal_soft_deprecated(paste_line(
+    "`reduce_right()` is soft-deprecated as of purrr 0.3.0.",
+    "Please use the new `.dir` argument of `reduce()` instead.",
+    "",
+    "  # Before:",
+    "  reduce_right(1:3, f)",
+    "",
+    "  # After:",
+    "  reduce(1:3, f, .dir = \"right\")  # New algorithm",
+    "  reduce(rev(1:3), f)             # Same algorithm as reduce_right()",
+    ""
+  ))
+  .x <- rev(.x) # Compatibility
+  reduce_impl(.x, .f, ..., .init = .init)
 }